+++ arduino_library/Plotter/Plotter.h
@@ -11,24 +11,26 @@ class Plotter{
         void begin(uint32_t baud_rate);
 
         //Adding a new buffer
-        int add_line(int* x_buf, int* y_buf, int buf_size);
+        uint8_t add_line(int32_t* x_buf, int32_t* y_buf, uint8_t buf_size);
 
         //Managing buffers
-        bool push_to_buffer(int line_d, int x, int y);
-        void send_buffer(int line_d);
+        bool push_to_buffer(uint8_t line_d, int32_t x, int32_t y);
+        void send_buffer(uint8_t line_d);
 
         //should only be used as a helper function from send_buffer for now
-        void send_buffer_compact(int* x, int* y, int line_d, int num);
+        void send_buffer_compact(int32_t* x, int32_t* y, uint8_t line_d, uint8_t num);
 
         //lowkey depracated stuff
-        void send_point(int x, int y, int line_d);
+        void send_point(int32_t x, int32_t y, uint8_t line_d);
         
     private:
-        int next_line_d;
-        int* x_bufs[MAX_LINES];
-        int* y_bufs[MAX_LINES];
-        int idx_ptr[MAX_LINES];
-        int buf_sizes[MAX_LINES];
+        uint8_t next_line_d;
+        int32_t* x_bufs[MAX_LINES];
+        int32_t* y_bufs[MAX_LINES];
+
+        //Capping buffer sizes at 256 here. Might want to make larger later.
+        uint8_t idx_ptr[MAX_LINES];
+        uint8_t buf_sizes[MAX_LINES];
 };
 
 #endif
